[tempsave] 컴포넌트 분리 진행중 #127
